+++ packages/fkweb/api/urls.py
@@ -5,7 +5,7 @@ from django.conf.urls import url
 from rest_framework.routers import SimpleRouter
 from rest_framework.urlpatterns import format_suffix_patterns
 
-from fkws import views
+from . import views
 
 
 router = SimpleRouter()
