--- packages/fkweb/create/apps.py
@@ -1,7 +1,0 @@
-
-
-from django.apps import AppConfig
-
-
-class CreateConfig(AppConfig):
-    name = 'create'
