--- packages/fkweb/create/models.py
@@ -1,5 +1,0 @@
-
-
-from django.db import models
-
-# Create your models here.
